move poi api methods to vector layer add basic comments for terrain/image factory api methods
