+++ sdkproject/Assets/Mapbox/Unity/Map/AbstractMap.cs
@@ -1021,151 +1021,5 @@ namespace Mapbox.Unity.Map
 			_options.scalingOptions.HasChanged = true;
 		}
 		#endregion
-
-		#region Location Prefabs Methods
-
-		/// <summary>
-		/// Places a prefab at the specified LatLon on the Map.
-		/// </summary>
-		/// <param name="prefab"> A Game Object Prefab.</param>
-		/// <param name="LatLon">A Vector2d(Latitude Longitude) object</param>
-		public void SpawnPrefabAtGeoLocation(GameObject prefab,
-											 Vector2d LatLon,
-											 Action<List<GameObject>> callback = null,
-											 bool scaleDownWithWorld = true,
-											 string locationItemName = "New Location")
-		{
-			var latLonArray = new Vector2d[] { LatLon };
-			SpawnPrefabAtGeoLocation(prefab, latLonArray, callback, scaleDownWithWorld, locationItemName);
-		}
-
-		/// <summary>
-		/// Places a prefab at all locations specified by the LatLon array.
-		/// </summary>
-		/// <param name="prefab"> A Game Object Prefab.</param>
-		/// <param name="LatLon">A Vector2d(Latitude Longitude) object</param>
-		public void SpawnPrefabAtGeoLocation(GameObject prefab,
-											 Vector2d[] LatLon,
-											 Action<List<GameObject>> callback = null,
-											 bool scaleDownWithWorld = true,
-											 string locationItemName = "New Location")
-		{
-			var coordinateArray = new string[LatLon.Length];
-			for (int i = 0; i < LatLon.Length; i++)
-			{
-				coordinateArray[i] = LatLon[i].x + ", " + LatLon[i].y;
-			}
-
-			PrefabItemOptions item = new PrefabItemOptions()
-			{
-				findByType = LocationPrefabFindBy.AddressOrLatLon,
-				prefabItemName = locationItemName,
-				spawnPrefabOptions = new SpawnPrefabOptions()
-				{
-					prefab = prefab,
-					scaleDownWithWorld = scaleDownWithWorld
-				},
-
-				coordinates = coordinateArray
-			};
-
-			if (callback != null)
-			{
-				item.OnAllPrefabsInstantiated += callback;
-			}
-
-			CreatePrefabLayer(item);
-		}
-
-		/// <summary>
-		/// Places the prefab for supplied categories.
-		/// </summary>
-		/// <param name="prefab">GameObject Prefab</param>
-		/// <param name="categories"><see cref="LocationPrefabCategories"/> For more than one category separate them by pipe
-		/// (eg: LocationPrefabCategories.Food | LocationPrefabCategories.Nightlife)</param>
-		/// <param name="density">Density controls the number of POIs on the map.(Integer value between 1 and 30)</param>
-		/// <param name="locationItemName">Name of this location prefab item for future reference</param>
-		/// <param name="scaleDownWithWorld">Should the prefab scale up/down along with the map game object?</param>
-		public void SpawnPrefabByCategory(GameObject prefab,
-										  LocationPrefabCategories categories = LocationPrefabCategories.AnyCategory,
-										  int density = 30, Action<List<GameObject>> callback = null,
-										  bool scaleDownWithWorld = true,
-										  string locationItemName = "New Location")
-		{
-			PrefabItemOptions item = new PrefabItemOptions()
-			{
-				findByType = LocationPrefabFindBy.MapboxCategory,
-				categories = categories,
-				density = density,
-				prefabItemName = locationItemName,
-				spawnPrefabOptions = new SpawnPrefabOptions()
-				{
-					prefab = prefab,
-					scaleDownWithWorld = scaleDownWithWorld
-				}
-			};
-
-			if (callback != null)
-			{
-				item.OnAllPrefabsInstantiated += callback;
-			}
-
-			CreatePrefabLayer(item);
-		}
-
-		/// <summary>
-		/// Places the prefab at POI locations if its name contains the supplied string
-		/// <param name="prefab">GameObject Prefab</param>
-		/// <param name="nameString">This is the string that will be checked against the POI name to see if is contained in it, and ony those POIs will be spawned</param>
-		/// <param name="density">Density (Integer value between 1 and 30)</param>
-		/// <param name="locationItemName">Name of this location prefab item for future reference</param>
-		/// <param name="scaleDownWithWorld">Should the prefab scale up/down along with the map game object?</param>
-		/// </summary>
-		public void SpawnPrefabByName(GameObject prefab,
-									  string nameString,
-									  int density = 30,
-									  Action<List<GameObject>> callback = null,
-									  bool scaleDownWithWorld = true,
-									  string locationItemName = "New Location")
-		{
-			PrefabItemOptions item = new PrefabItemOptions()
-			{
-				findByType = LocationPrefabFindBy.POIName,
-				nameString = nameString,
-				density = density,
-				prefabItemName = locationItemName,
-				spawnPrefabOptions = new SpawnPrefabOptions()
-				{
-					prefab = prefab,
-					scaleDownWithWorld = scaleDownWithWorld
-				}
-			};
-
-			CreatePrefabLayer(item);
-		}
-
-		/// <summary>
-		/// Creates the prefab layer.
-		/// </summary>
-		/// <param name="item"> the options of the prefab layer.</param>
-		private void CreatePrefabLayer(PrefabItemOptions item)
-		{
-			if (_vectorData.LayerProperty.sourceType == VectorSourceType.None
-			|| !_vectorData.LayerProperty.sourceOptions.Id.Contains(MapboxDefaultVector.GetParameters(VectorSourceType.MapboxStreets).Id))
-			{
-				Debug.LogError("In order to place location prefabs please add \"mapbox.mapbox-streets-v7\" to the list of vector data sources");
-				return;
-			}
-
-			//ensure that there is a vector layer
-			if (_vectorData == null)
-			{
-				_vectorData = new VectorLayer();
-			}
-
-			_vectorData.AddLocationPrefabItem(item);
-		}
-
-		#endregion
 	}
 }
