+++ sdkproject/Assets/Mapbox/Unity/SourceLayers/TerrainLayer.cs
@@ -217,6 +217,10 @@
 
 		#region API Methods
 
+		/// <summary>
+		/// Set the data source for the terrain layer. Can be used to disable terrain layer by setting it "None".
+		/// </summary>
+		/// <param name="dataSource"></param>
 		public void SetDataSource(ElevationSourceType dataSource)
 		{
 			if (_layerProperty.sourceType != dataSource)
@@ -226,6 +230,11 @@
 			}
 		}
 
+		/// <summary>
+		/// Sets the main strategy for terrain mesh generation.
+		/// Flat terrain doesn't pull data from servers and just uses a quad as terrain.
+		/// </summary>
+		/// <param name="elecationType">Type of the elevation strategy</param>
 		public void SetElevationType(ElevationLayerType elecationType)
 		{
 			if (LayerProperty.elevationLayerType != elecationType)
@@ -235,6 +244,10 @@
 			}
 		}
 
+		/// <summary>
+		/// Add/Remove terrain collider. Terrain uses mesh collider.
+		/// </summary>
+		/// <param name="enable">Boolean for enabling/disabling mesh collider</param>
 		public void EnableCollider(bool enable)
 		{
 			if (LayerProperty.colliderOptions.addCollider != enable)
@@ -244,6 +257,10 @@
 			}
 		}
 
+		/// <summary>
+		/// Sets the elevation multiplier for terrain. It'll regenerate terrain mesh, multiplying each point elevation by provided value.
+		/// </summary>
+		/// <param name="factor">Elevation multiplier</param>
 		public void SetExaggerationFactor(float factor)
 		{
 			if (LayerProperty.requiredOptions.exaggerationFactor != factor)
